Fix #707, Resolve highest MsgID of 0xFFFF bug
Changes Message Key from uint16 to uint32 to avoid rollover

+++ fsw/cfe-core/src/sb/cfe_sb_priv.h
@@ -119,7 +119,7 @@
 **           be used directly, except by internal table lookups.
 **
 */
-typedef uint16  CFE_SB_MsgKey_Atom_t;
+typedef uint32  CFE_SB_MsgKey_Atom_t;
 
 /******************************************************************************
 **  Typedef:  CFE_SB_MsgKey_t
